Module EzSession.Hash

Hash module, that hashing algorithms.

include module type of struct include EzHash end

val hash : string -> string

val password : login:string -> password:string -> string

Hashed version of the password that is computed by the hash function applied on login ^ password

val challenge : challenge:string -> pwhash:string -> string

Hashed version of the challenge that is computed by the hash function applied on challenge ^ pwhash