Module EzSession.Make

Main functor that produces definition for authentication services and encodings for types used by service's input, output and errors.

Parameters

module S : TYPES.SessionArg

Signature

type nonrec auth = (S.user_id, S.user_info) TYPES.auth

module Encoding : sig ... end

Encodings for data types used in server's requests/responses and for error cases that could be raised by one of them.

module Service : sig ... end

Definition for services and their security's configuration.