Module Make.Service

Definition for services and their security's configuration.

type user_id = S.user_id

type user_info = S.user_info

type nonrec auth = auth

val section_session : EzAPI.Doc.section

Documentation section for openapi.

val param_token : EzAPI.Param.t

Parameter with name token that stores an authentication token string

type token_security = [

| EzAPI.Security.cookie

| EzAPI.Security.header

| EzAPI.Security.query

]

Type that represents security by authentication token and the way that request uses to store it.

val param_security : [> `Query of EzAPI.Param.t EzAPI.Security.apikey ]

Security that requires param_token parameter in query.

val header_cookie_security : [> `Cookie of string EzAPI.Security.apikey * int64 option | `Header of string EzAPI.Security.apikey ]

Security that checks S.token_kind: If it is a CSRF token, then requires a CSRF header. Otherwise requires token to be found in the cookies.

val security : token_security list

Security that combines param_security and header_cookie_security in the corresponding order. Represents the security configuration for connect and logout requests.

val rpc_root : (EzAPI.Req.t, EzAPI.Req.t) EzAPI.Path.t

Defines path to authentication services

val headers : EzAPI.StringSet.t EzAPI.StringMap.t

val connect : (auth TYPES.connect_response, TYPES.connect_error, token_security) EzAPI.service0

Connection service that requires authentication token. For more details, see corresponding EzSessionServer.Make.connect handler and default client request implementation

EzSessionClient.Make.connect.

val login : (TYPES.login_message, (S.user_id, S.user_info) TYPES.login_response, TYPES.login_error, EzAPI.Security.none) EzAPI.post_service0

Logining service. For more details, see corresponding EzSessionServer.Make.login handler and default client request implementation EzSessionClient.Make.login.

val logout : (TYPES.auth_needed, TYPES.logout_error, token_security) EzAPI.service0

Disconnection service that requires authentication token. For more details, see corresponding EzSessionServer.Make.logout handler and default client request implementation EzSessionClient.Make.logout.