Module EzSession.TYPES

type foreign_info = {

foreign_origin : string;

foreign_token : string;

}

Foreign user without password, whose login is equal to foreign_origin ^ foreign_token and whose password is not req.

type 'user_id session = {

session_token : string;

session_login : string;

session_user_id : 'user_id;

session_last : float;

session_foreign : foreign_info option;

}

A session that helps to keep connection for the given user and stores useful information about his communication with the sever.

module type SessionArg = sig ... end

Main module that specifies the session parameters for given implementation of API server.

type ('user_id, 'user_info) auth = {

auth_login : string;

auth_user_id : 'user_id;

auth_token : string;

auth_user_info : 'user_info;

}

Authentification information returned by server after successful connection

type auth_needed = {

challenge_id : string;

challenge : string;

}

Challenge that should be resolved to be able to connect

type 'auth connect_response =

| AuthOk of 'auth

| AuthNeeded of auth_needed

Connection response, that either describes user information if connection successes either describes challenge to resolve to be able to connect

type local_login_message = {

login_user : string;

login_challenge_id : string;

login_challenge_reply : string;

}

Logining request, that contains the login and the challenge resolution obtained by hashing challenge and password provided by the user.

type login_message =

| Local of local_login_message

| Foreign of foreign_info

Logining request, that could be asked either by user with the password provided either by a foreign user without a password.

type ('user_id, 'user_info) login_response =

| LoginOk of ('user_id, 'user_info) auth

| LoginWait of 'user_id

Possible logining outcomes.

type login_error = [

| `Bad_user_or_password

| `User_not_registered

| `Unverified_user

| `Challenge_not_found_or_expired of string

| `Invalid_session_login of string

]

Errors that could be raised while logining.

type logout_error = [

| `Invalid_session_logout of string

]

Errors that could be raised while disconnecting.

type connect_error = [

| `Session_expired

| `Invalid_session_connect of string

]

Errors that could be raised while connecting.