Module EzSessionServer

val random_challenge : unit -> string

Returns random challenge 30 characters long

module type SessionStore = sig ... end

Module that defines a way to store user's sessions.

module type Arg = sig ... end

Main argument for Make functor. Specifies session's and user's storage and session main configurations.

val default_check_foreign : origin:string -> token:string -> (_, int * string option) Stdlib.result Lwt.t

Not implemented search for foreign user.

val default_register_foreign : origin:string -> token:string -> (_, int * string option) Stdlib.result Lwt.t

Not implemented registration for foreign user.

module Make (S : Arg) : sig ... end

Main functor that defines handlers for authentication services.

exception UserAlreadyDefined

Exception when trying to create user that already exists.

exception NoPasswordProvided

Exception when trying to create non-foreign user without password.

module UserStoreInMemory (S : EzSession.TYPES.SessionArg with type user_id = string) : sig ... end

User storage located in the memory of the server.

module SessionStoreInMemory : SessionStore

Implementation of session store in the server's memory.