Parameter Make.S

module SessionArg : EzSession.TYPES.SessionArg

See EzSession.TYPES.SessionArg

module SessionStore : SessionStore with type user_id = SessionArg.user_id

See SessionStore

val find_user : login:string -> (string option * SessionArg.user_id * SessionArg.user_info) option Lwt.t

Searches for user with given login and returns its password, if defined, its id and its information.

val check_foreign : origin:string -> token:string -> (string, int * string option) Stdlib.result Lwt.t

Checks for foreign user. If exists, returns its login. Otherwise returns an error number with an error message.

val register_foreign : origin:string -> token:string -> (SessionArg.user_id * SessionArg.user_info option, int * string option) Stdlib.result Lwt.t

Registers foreign user.