Module S.SessionStore

See SessionStore

type user_id = SessionArg.user_id

User id. See EzSession.TYPES.SessionArg.user_id.

val create_session : ?foreign:EzSession.TYPES.foreign_info -> login:string -> req:EzAPIServerUtils.Req.t -> user_id -> user_id EzSession.TYPES.session Lwt.t

Creates session from the specified information, generates random token, and associates session to the given token in storage.

val get_session : ?req:EzAPIServerUtils.Req.t -> string -> user_id EzSession.TYPES.session option Lwt.t

Looks up for the session associated with the given token. If req is also specified, then the returned session synchronizes its time with request.

val remove_session : user_id -> token:string -> unit Lwt.t

Removes session from session_by_token if user_id corresponds to session's one.