Parameter UserStoreInMemory.S

type user_id = string

User identifier, which generally is the same as login

type user_info

Associated to user information

val web_host : string option

Web host, that should be used in access control headers, if specified. If web_host isn't specified, then acces-control header in response will be set to '*' and authentication wwith cookies wouldn't work. Note : Cookies would be set by browser only if request's flag 'with_credentials' is set to true. Last one in turn, requires that "access-control-allow-origin" header by reponse returns something different from "*".

val user_id_encoding : user_id Json_encoding.encoding

Json encoding for user's id

val user_info_encoding : user_info Json_encoding.encoding

Json encoding for user's information

val rpc_path : string list

RPC path where authentication services like connect, login and logout would be placed.

val token_kind : [ `Cookie of string * int64 option | `CSRF of string ]

Describes two ways to store a token within a client request :

Stores as a cookie associated with the given cookie name and its max-age, if provided.

Stores as a CSRF header with the given name.